// src/types/banner-json.ts
export type JsonGradColor = {
  c: string
  p: number
}
// ...
export type EditorLayerFeed = Record<string, string | null>
export type AdStudioLayerFeedPropertyBind = {
  hash: string
  column: string
  index?: string
} | null
export type AdStudioLayerFeed = Record<
  string,
  AdStudioLayerFeedPropertyBind | AdStudioLayerFeedPropertyBind[]
>

export type JsonPropertiesFeed = EditorLayerFeed | AdStudioLayerFeed
// ...
export type JsonRGradPos =
  | 'center center'
  | 'center right'
  | 'center left'
  | 'bottom left'
  | 'bottom center'
  | 'bottom right'
  | 'top left'
  | 'top center'
  | 'top right'

export type JsonAlign =
  | 'bottom'
  | 'center'
  | 'left'
  | 'middle'
  | 'right'
  | 'top'

export type JsonScaleMode =
  | 'stretch'
  | 'aspect'
  | 'mask'
  | 'tile'
  | 'crop'
  | 'userCrop'

export type JsonBackgroundType = 'solid' | 'rgrad' | 'lgrad' | 'image' | 'none'

export type JsonBackgroundSolid = {
  type: 'solid'
  scolor: string
}

export type JsonBackgroundSolidText = JsonBackgroundSolid & {
  feed?: AdStudioLayerFeedPropertyBind
}

export type JsonBackgroundLinearGradient = {
  type: 'lgrad'
  gradColors: JsonGradColor[]
  backgroundRotation: number
}

export type JsonBackgroundRadialGradient = {
  type: 'rgrad'
  gradColors: JsonGradColor[]
  rgradPos: JsonRGradPos
}

export type JsonBackgroundImage = {
  type: 'image'
  originalWidth: number
  originalHeight: number
  url: string
  backgroundRotation?: number
  scaleMode: JsonScaleMode
  contentScale?: number
  horizontalAlign: JsonAlign
  verticalAlign: JsonAlign
  contentOffsetX?: number
  contentOffsetY?: number
  hqUrl?: string
  source?: string
  hash?: string
  licensing?: boolean
}

export type JsonBackgroundNone = {
  type: 'none'
}

export interface JsonBackgroundBorder {
  useBorder?: boolean
  borderColor?: string
}

export type JsonBackground =
  | JsonBackgroundNone
  | JsonBackgroundSolid
  | JsonBackgroundLinearGradient
  | JsonBackgroundRadialGradient
  | JsonBackgroundImage

export type JsonBackgroundWithBorder = JsonBackgroundBorder & JsonBackground

export type JsonTextGradient =
  | JsonBackgroundNone
  | JsonBackgroundSolidText
  | JsonBackgroundLinearGradient
  | JsonBackgroundRadialGradient
  | JsonBackgroundImage
// ...
export type JsonTransitionType =
  | 'none'
  | 'instant'
  | 'alpha'
  | 'alpha-words'
  | 'blur'
  | 'blur-words'
  | 'slide'
  | 'slideBounce'
  | 'slideElastic'
  | 'roll'
  | 'zoom'
  | 'flip'
  | 'rotate'
  | 'scale'
  | 'show'
  | 'hide'
  | 'cross'
  | 'scaleBounce'
  | 'scaleElastic'
  | 'vibrate'
  | 'flicker'
  | 'shake'
  | 'jello'
  | 'bounce'
  | 'pulsate'
  | 'blink'
  | 'video'
  | 'audio'
  | 'advanced'
  | 'shadowPop'
  | 'fadeAudio'
  | 'swing'
  | 'fly'
  | 'tilt'
  | 'slit'
  | 'step'
  | 'kenBurns'

export type JsonTransitionTypeSlide =
  | 'none'
  | 'alpha'
  | 'blur'
  | 'slide'
  | 'slideBounce'
  | 'slideElastic'
  | 'scale'
  | 'scaleElastic'
  | 'scaleBounce'
  | 'roll'
  | 'zoom'
  | 'flip'
  | 'rotate'
  | 'swing'
  | 'fly'
  | 'tilt'
  | 'slit'
  | 'step'

export type JsonTransitionCrossType = 'cross' | 'hide' | 'show'

export type JsonTransitionDirection =
  | 'custom'
  | 'l2r'
  | 'r2l'
  | 't2b'
  | 'b2t'
  | 'top'
  | 'left'
  | 'right'
  | 'bottom'
  | 'topRight'
  | 'topLeft'
  | 'bottomRight'
  | 'bottomLeft'
  | 'forward'
  | 'backward'
  | 'center'
  | 'vertical'
  | 'horizontal'
  | 'random'
  | 'topForward'
  | 'topBack'
  | 'rightForward'
  | 'rightBack'
  | 'bottomForward'
  | 'bottomBack'
  | 'leftForward'
  | 'leftBack'

export type JsonSlideTransition = JsonTransition & {
  type: JsonTransitionTypeSlide
  crosstype: JsonTransitionCrossType
}

export type JsonTweenType =
  | 'Sine'
  | 'Quad'
  | 'Cubic'
  | 'Quart'
  | 'Quint'
  | 'Expo'
  | 'Circ'
  | 'Strong'
  | 'Bounce'
  | 'Back'
  | 'Elastic'

export type JsonEaseType =
  | 'linear'
  | 'ease'
  | 'easeIn'
  | 'easeOut'
  | 'easeInOut'
// ...
export type AdvancedProperties = {
  [AdvancedPanelItems.MOVE]: AdvancedMoveProperty[]
  [AdvancedPanelItems.ROTATE]: AdvancedRotateProperty[]
  [AdvancedPanelItems.SCALE]: AdvancedScaleProperty[]
  [AdvancedPanelItems.OPACITY]: AdvancedOpacityProperty[]
  [AdvancedPanelItems.BLUR]: AdvancedBlurProperty[]
}

export type JsonTransition = {
  type: JsonTransitionType
  direction: JsonTransitionDirection
  duration: number
  delay: number
  tween: JsonTweenType
  ease: JsonEaseType
  wordsDuration?: number
  wordsAppearOrder?: 'l2r' | 'r2l'
  slidePosX: number
  slidePosY: number
  alphaOffset: number
  slideOffset: number
  blurAmount: number
  zoom: number
  preset: string
  color?: string
  shadowLength?: number
  advancedProperties?: AdvancedProperties
}

export type JsonTransitionGradColor = {
  gradColors: JsonGradColor[]
}

export type JsonTransitionMid = {
  type: JsonTransitionType
  direction: 'vertical' | 'horizontal' | 'top'
  variation?: string | null
  duration: number
  delay: number
  motionTime: number
  ease: JsonEaseType
  tween: JsonTweenType
  preset: string
  startEnd: string
  advancedProperties?: AdvancedProperties
  gradient?: JsonTransitionGradColor
}
// ...
export type JsonShadow = {
  useShadow: boolean
  hShadow: number
  vShadow: number
  blur: number
  spread: number
  color: string
}

export type JsonBlur = {
  useBlur: boolean
  pixels: number
}

export const JsonFlipTypeArr = [
  'both',
  'horizontal',
  'none',
  'vertical',
] as const

export type JsonFlipType = (typeof JsonFlipTypeArr)[number]

export type JsonTextAlignment = 'left' | 'right' | 'center' | 'justify'

export const JsonTextVerticalAlignment = {
  top: 'top',
  center: 'center',
  bottom: 'bottom',
} as const
export type JsonTextVerticalAlignmentType =
  (typeof JsonTextVerticalAlignment)[keyof typeof JsonTextVerticalAlignment]
// ...
export const JsonTextContentHeight = {
  auto: 'auto',
  fixed: 'fixed',
} as const
export type JsonTextContentHeightType =
  (typeof JsonTextContentHeight)[keyof typeof JsonTextContentHeight]

export type JsonActionGotoURL = {
  event: string
  target: JsonActionTarget
  type: 'gotoURL'
  url: string
  useHandCursor: boolean
}

export type JsonActionGotoSlide = {
  event: string
  type: 'gotoSlide'
  slide: JsonActionSlide
  useHandCursor: boolean
}

export type JsonActionVideoEnd = {
  event: string
  target: string | null
  type: 'gotoSlide'
  url: string
  slide: JsonActionSlide
  useHandCursor: boolean
}

export type JsonActionNone = {
  type: 'none'
  event: string
}

export type JsonAction =
  | JsonActionGotoURL
  | JsonActionGotoSlide
  | JsonActionNone
  | JsonActionVideoEnd

export function isJsonActionGotoSlide(
  action: JsonAction
): action is JsonActionGotoSlide {
  return action.type === 'gotoSlide'
}

export function isJsonActionGotoURL(
  action: JsonAction
): action is JsonActionGotoURL {
  return action.type === 'gotoURL'
}

export function isJsonActionNone(action: JsonAction): action is JsonActionNone {
  return action.type === 'none'
}

export type JsonBaseElementProperties = {
  id: number
  bannersetElementId?: number | null
  elementCategory?: string | null
  layerName: string
  dropShadow?: JsonShadow
  blur?: JsonBlur
  opacity: number
  rotation: number
  flip?: JsonFlipType
  x: number
  y: number
  width: number
  height: number
  buildIn?: JsonTransition
  buildMid?: JsonTransitionMid
  buildOut?: JsonTransition
  actions?: JsonAction[]
  locked?: boolean
  lockedProperties?: string[]
  visible: boolean
  group?: number | null
  feed?: JsonPropertiesFeed
}

export type JsonTextProperties = JsonBaseElementProperties & {
  text?: string
  alignment: JsonTextAlignment
  verticalAlign?: JsonTextVerticalAlignmentType
  contentHeightType: JsonTextContentHeightType
  fontSize: number
  lineHeight: number
  letterSpacing: number
  config: JsonTextSlateConfig // TODO:
  textShadow?: JsonShadow
  initialFontSize?: number
  textDirection?: string | null
  scale?: number
  blendMode?: JsonBlendMode
  outline?: JsonTextOutline
}

export type JsonTextOutline = {
  useOutline?: boolean
  color?: string
  weight?: number
}

export type JsonText = {
  type: 'layer'
  layerType: 'text'
  properties: JsonTextProperties
}

export type JsonOldTextV2 = {
  type: 'layer'
  layerType: 'text'
  properties: JsonTextProperties & { config: JsonTextOldConfig[] }
}

export type JsonOldTextV1 = {
  properties: JsonTextOldConfigStyle
}

// TODO: add all attributes
export type JsonTextSlateConfigNode = {
  defaultFontSettings?: JsonFontSettings
  children: Array<JsonTextSlateConfigChildren | JsonTextSlateConfigNode>
  type:
    | 'paragraph'
    | 'legacy'
    | 'subparagraph'
    | 'wrapper'
    | 'bullet-list'
    | 'number-list'
    | 'list-item'
    | 'feed-chip'
}
// ...
export type JsonFontSettings = {
  fontFamily: string
  fontType: FontType
  fontStyle: FontStyle
  fontPrefix: string | null
  fontWeight: number
  fontFaceUrl: string | null
  fontUrl: string | null
}

export type JsonTextSlateConfigChildren = {
  color: string
  textScript: 'none' | 'superscript' | 'subscript'
  textDecoration: TextDecorationEnum
  textTransform: TextTransform
  fontSettings: JsonFontSettings
  typography: 'heading' | 'subheading' | 'body'
  gradient?: JsonTextGradient
  fontSizePercent: number
  fontSize?: number
  text: string
  feed?: AdStudioLayerFeedPropertyBind
  feedText?: AdStudioLayerFeedPropertyBind
}

export enum TextContentBoundaryType {
  OVERFLOW = 'overflow',
  SHRINK = 'shrink',
  ELLIPSIS = 'ellipsis',
}

export type Boundary = {
  type?: JsonTextContentBoundaryType
  minFontSize?: number
}

export type JsonTextSlateConfig = {
  type: 'slate'
  nodes: JsonTextSlateConfigNode[]
  selection?: unknown
  keepSelection?: boolean
  boundary?: Boundary
}
// ...
export type JsonBorder = {
  color: string
  weight: number
  radius: number
}

export type FontType = 'google' | 'custom' | 'all' | 'bannersnack' | 'creatopy'
export type FontStyle = 'normal' | 'italic' | 'regular' | 'slanted' | 'oblique'
export type TextTransform = 'none' | 'uppercase' | 'lowercase' | 'capitalize'

export type JsonButtonLabelStyle = JsonFontSettings & {
  fontDesign?: string
  textTransform?: TextTransform
  fontSize?: number
  color?: string
  letterSpacing?: number
  dropShadow?: JsonShadow
  initialFontSize?: number
  textDirection?: string
  fontPrefix?: string | null
}
// ...
export type JsonAdjustColor = {
  useAdjustColor: boolean
  brightness: number
  contrast: number
  saturate: number
  hue: number
}

export type JsonCropData = {
  x: number
  y: number
  width: number
  height: number
}

export type JsonImageProperties = JsonBaseElementProperties & {
  url: string
  scaleMode: JsonScaleMode
  horizontalAlign: JsonAlign
  verticalAlign: JsonAlign
  adjustColor: JsonAdjustColor
  contentScale: number
  contentOffsetX: number
  contentOffsetY: number
  originalWidth: number
  originalHeight: number
  hqUrl?: string
  cropData?: JsonCropData
  source?: string
  hash?: string
  licensing?: boolean
  blendMode?: JsonBlendMode
  mediaFolderId?: number
  originalName?: string
  maskImage?: string
}
// ...
export type JsonBlendMode =
  | 'color'
  | 'color-burn'
  | 'color-dodge'
  | 'darken'
  | 'difference'
  | 'exclusion'
  | 'hard-light'
  | 'hue'
  | 'lighten'
  | 'luminosity'
  | 'multiply'
  | 'normal'
  | 'overlay'
  | 'saturation'
  | 'screen'
  | 'soft-light'
// ...
export enum TextDecorationEnum {
  NONE = 'none',
  UNDERLINE = 'underline',
  STRIKETHROUGH = 'line-through',
}